requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
kombu==5.5.4
numpy==2.3.2
packaging==25.0
pandas==2.3.2
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
